requirements.txt
# Deep learning frameworks
torch>=1.10.0
pytorch-lightning>=1.5.0

# Transformers for models
transformers>=4.15.0

# Image processing
Pillow>=8.0.0

# Data processing
pandas>=1.3.0
numpy>=1.20.0
scikit-learn>=1.0.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# File handling and utilities
pathlib
requests>=2.25.0
tensorboard>=2.7.0